pyproject.toml
[project]
name = "foundry-wrap"
dynamic = ["version"]
description = "A Python wrapper for Foundry's forge scripts with dynamic interface generation"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "click>=8.1.0",
    "toml>=0.10.2",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "safe-eth-py>=7.0.0",
    "hatch>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0"
]
requires-python = ">=3.8"

[project.scripts]
foundry-wrap = "foundry_wrap.cli:main"

[build-system]
requires = ["hatchling>=1.13.0", "pathspec>=0.10.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foundry_wrap"]

[tool.hatch.version]
path = "src/foundry_wrap/version.py"